refactor(orm)!: consolidate query.ts into filter.ts + query-options.ts
RALPH: task=issue-113, prd=issue-112

## What
Extract `OrderBy` and `QueryOptions<Sel>` from the legacy
`src/orm/query.ts` into a new `src/orm/query-options.ts` Module.
Hard-delete `query.ts` — all legacy filter vocabulary (`Where`,
`WhereOp`, `WhereGroupOp`, `QueryGroup`, `WhereFactory`, `FilterOp`)
is removed from the public API. The canonical `filter.ts` already
carries full test coverage for the equivalent `FilterGroup` behavior
(clone-on-step, fan-out independence, empty-combinator rejection,
schema-field-ref).

## Key decisions
- Hard-remove with no transitional shim — the §4.1 name-parity
  violations (`nin`/`nexists`/`ncontains`) cannot be kept alive
- `OrderBy` + `QueryOptions` get their own Module because they are
  orthogonal to filter trees (filter = which rows; options = how to
  slice them)
- All inline tests from `query.ts` already had equivalent coverage
  in `filter.ts` — no test lift needed, just deletion

## Files changed
- src/orm/query-options.ts (new) — `OrderBy` class + `QueryOptions<Sel>` type
- src/orm/query.ts (deleted) — legacy filter module (208 LOC)
- src/orm/utils.ts — `AnyOrmClass` union: `Where`/`QueryGroup` → `Filter`/`FilterGroup`
- src/orm/orm-adapter.ts — `QueryOptions` import path → `./query-options`
- src/orm/adapters/base.ts — `QueryOptions` import path → `../query-options`
- src/orm/adapters/{in-memory,json,mongodb,postgresql}/ — import paths → `../../query-options`
- src/orm/repo/builders.ts — `OrderBy` import path → `../query-options`
- src/orm/repo/internals/executors.ts — `OrderBy` import path → `../../query-options`
- src/orm/index.ts — barrel: `./query` → `./query-options`
- docs/orm/CONTEXT.md — §11.1 + §3.1 notes for `query-options.ts`

## Blockers/notes
- Net LOC: -208 deleted, +18 added = -190
- 648 tests pass, typecheck clean
- Adapter-internal `./query` modules (mongodb, postgresql) are
  unaffected — they reference their own local query compilers

BREAKING CHANGE: `Where`, `WhereOp`, `WhereGroupOp`, `QueryGroup`,
`WhereFactory`, and `FilterOp` are removed from the public API.
Use `Filter`, `FilterGroup`, and `FilterFactory` from `filter.ts`.
