+++ src/orm/adapters/base.ts
@@ -1,6 +1,6 @@
 import type { FilterGroup } from '../filter'
 import type { AggregateSpec } from '../orm-adapter'
-import type { QueryOptions } from '../query'
+import type { QueryOptions } from '../query-options'
 import type { AnySchema } from '../schema'
 import type { AnyUpdateOp } from '../updates'
