+++ src/orm/adapters/postgresql/query.ts
@@ -2,7 +2,7 @@ import type { AggregateOpName } from '../../adapter'
 import { OrmValidationError } from '../../errors'
 import { Filter, FilterGroup, type FilterChild } from '../../filter'
 import type { AggregateSpec } from '../../orm-adapter'
-import type { QueryOptions } from '../../query'
+import type { QueryOptions } from '../../query-options'
 import { IncOp, MaxOp, MinOp, MulOp, PatchOp, PullOp, PushOp, UnsetOp } from '../../updates'
 
 function mapField(field: string, primaryKey: string): string {
@@ -350,7 +350,7 @@ export function buildAggregateQuery(
 if (import.meta.vitest) {
 	const { describe, test, expect } = import.meta.vitest
 	const { FilterGroup } = await import('../../filter')
-	const { OrderBy } = await import('../../query')
+	const { OrderBy } = await import('../../query-options')
 	const { IncOp, MulOp, MinOp, MaxOp, UnsetOp, PushOp, PullOp, PatchOp, SetOp, flattenOps } = await import('../../updates')
 	const { OrmValidationError } = await import('../../errors')
