+++ src/orm/index.ts
@@ -5,7 +5,7 @@ export * from './errors'
 export * from './event-log'
 export * from './fields'
 export * from './filter'
-export * from './query'
+export * from './query-options'
 export * from './relations'
 export * from './repo'
 export * from './schema'
