+++ src/orm/query-options.ts
@@ -1,0 +1,18 @@
+import { toFieldName, type AnyField } from './fields'
+
+export class OrderBy {
+	readonly field: string
+	constructor(
+		field: string | AnyField,
+		readonly direction: 'asc' | 'desc',
+	) {
+		this.field = toFieldName(field)
+	}
+}
+
+export type QueryOptions<Sel extends string = string> = {
+	orderBy?: OrderBy[]
+	limit?: number
+	offset?: number
+	select?: readonly Sel[]
+}
