+++ src/orm/repo/builders.ts
@@ -4,7 +4,7 @@ import { OrmNotFoundError, type OrmNotFoundOperation } from '../errors'
 import { toFieldName, type AnyField, type Field } from '../fields'
 import { FilterGroup, type FilterFactory } from '../filter'
 import type { AggregateSpec } from '../orm-adapter'
-import { OrderBy } from '../query'
+import { OrderBy } from '../query-options'
 import type { AnyPreloadDef } from '../relations'
 import type { AnySchema, SchemaOutput } from '../schema'
 import type { SchemaCreateInput, SchemaUpdateInput } from '../schema-validations'
