+++ src/orm/repo/internals/executors.ts
@@ -2,7 +2,7 @@ import { planSelection } from './computeds'
 import type { SelectedWithPreloads } from './types'
 import { assertNormalisedAggregate, assertNormalisedFilter, type FilterGroup } from '../../filter'
 import type { AggregateSpec } from '../../orm-adapter'
-import type { OrderBy } from '../../query'
+import type { OrderBy } from '../../query-options'
 import type { AnyPreloadDef } from '../../relations'
 import type { AnySchema } from '../../schema'
 import { validateCreate, validateCreateMany, validateUpdate, validateUpsertConflicts, type SchemaCreateInput, type SchemaUpdateInput } from '../../schema-validations'
